Updated the SpatialIndex files to use the new version of the Spatial Index library (v1.7.1). Thanks to Wasim Sabir
git-svn-id: https://mapwingis.svn.codeplex.com/svn@69521 142e9c35-602f-4880-9f37-ec48a24daa6a

+++ trunk/Utilities/SpatialIndex/IndexShapeFiles.cpp
@@ -193,7 +193,7 @@ namespace IndexSearching
 
 		// data should be an array of characters representing a Region as a string.
 		byte* pData = 0;
-		size_t cLen = 0;
+		uint32_t cLen = 0;
 		d.getData(cLen, &pData);
 		delete[] pData;
